implement slot
